+++ src/components/RobotMenu.jsx
@@ -1,6 +1,6 @@
 import { useSelector } from "react-redux";
 import GameButton from "./GameButton";
-import { buttonCall } from "../modules/engine";
+import { buttonCall } from "../modules/buttons";
 
 function RobotMenu() {
     const robotMenu = useSelector(store => store.robotMenu);
